chore: rename apis module to bedev2
